+++ src/components/TransferModal.js
@@ -221,7 +221,7 @@ const TransferModal = ({ account, wrapSetShow }) => {
                         paddingTop: '0.2rem',
                         backgroundColor: '#1f1f1f',
                         color: '#F6F3FB'
-                    }} min={0} max={parseInt(amount) / 1000000} step={0.000001} onChange={handleChange} />
+                    }} min={0} max={parseFloat(amount) / 1000000} step={0.000001} onChange={handleChange} />
                 </>
             </div>
             <div>
